refactor: scatter mmap.rs contents to unix/windows modules
This is some duplication, but it once and for all eliminates all cfg
fuckery and code that expects precisely one module to be defined.

Admittedly, for windows/unix we could have stuck with the cfg stuff, but
I wanted a clean slate for now. The windows parts are compile-tested
only, as I do not have a windows system available.

The handling of various from_range functions is based on the assumption
that these were only used in test code and nowhere else (hence all the
preceding refactors of test code). In unit tests, their use has been
eliminated, and doc tests now got a cfg(target_family="unix",
not(feature="xen")) so that they only run when mmap_unix is available.
This is fine, because they are examples of how to use specific
functions, and not any serious tests (e.g. dont need to run for all
supported backends).
