+++ src/bitmap/mod.rs
@@ -125,7 +125,7 @@ pub(crate) mod tests {
     use std::sync::atomic::Ordering;
 
     use crate::{Bytes, VolatileMemory};
-    #[cfg(feature = "backend-mmap")]
+    #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     use crate::{GuestAddress, MemoryRegionAddress};
 
     // Helper method to check whether a specified range is clean.
@@ -219,7 +219,7 @@ pub(crate) mod tests {
     // them to test the mmap-based backend implementations for now. Going forward, the generic
     // test functions defined here can be placed in a separate module (i.e. `test_utilities`)
     // which is gated by a feature and can be used for testing purposes by other crates as well.
-    #[cfg(feature = "backend-mmap")]
+    #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     fn test_guest_memory_region<R: GuestMemoryRegion>(region: &R) {
         let dirty_addr = MemoryRegionAddress(0x0);
         let val = 123u64;
@@ -253,7 +253,7 @@ pub(crate) mod tests {
         );
     }
 
-    #[cfg(feature = "backend-mmap")]
+    #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     // Assumptions about M generated by f ...
     pub fn test_guest_memory_and_region<M, F>(f: F)
     where
