+++ src/mmap/mod.rs
@@ -12,29 +12,17 @@
 //!
 //! This implementation is mmap-ing the memory of the guest into the current process.
 
-use std::borrow::Borrow;
-use std::ops::Deref;
-use std::result;
-
-use crate::address::Address;
-use crate::bitmap::{Bitmap, BS};
-use crate::guest_memory::{self, FileOffset, GuestAddress, GuestUsize, MemoryRegionAddress};
-use crate::region::{
-    GuestMemoryRegion, GuestMemoryRegionBytes, GuestRegionCollection, GuestRegionCollectionError,
-};
-use crate::volatile_memory::{VolatileMemory, VolatileSlice};
-
 // re-export for backward compat, as the trait used to be defined in mmap.rs
 pub use crate::bitmap::NewBitmap;
 
 #[cfg(all(not(feature = "xen"), target_family = "unix"))]
-mod unix;
+pub(super) mod unix;
 
 #[cfg(all(feature = "xen", target_family = "unix"))]
-pub(crate) mod xen;
+pub(super) mod xen;
 
 #[cfg(target_family = "windows")]
-mod windows;
+pub(super) mod windows;
 
 #[cfg(all(not(feature = "xen"), target_family = "unix"))]
 pub use unix::{Error as MmapRegionError, MmapRegion, MmapRegionBuilder};
@@ -47,183 +35,17 @@ pub use std::io::Error as MmapRegionError;
 #[cfg(target_family = "windows")]
 pub use windows::MmapRegion;
 
-/// [`GuestMemoryRegion`](trait.GuestMemoryRegion.html) implementation that mmaps the guest's
-/// memory region in the current process.
-///
-/// Represents a continuous region of the guest's physical memory that is backed by a mapping
-/// in the virtual address space of the calling process.
-#[derive(Debug)]
-pub struct GuestRegionMmap<B = ()> {
-    mapping: MmapRegion<B>,
-    guest_base: GuestAddress,
-}
-
-impl<B> Deref for GuestRegionMmap<B> {
-    type Target = MmapRegion<B>;
-
-    fn deref(&self) -> &MmapRegion<B> {
-        &self.mapping
-    }
-}
-
-impl<B: Bitmap> GuestRegionMmap<B> {
-    /// Create a new memory-mapped memory region for the guest's physical memory.
-    ///
-    /// Returns `None` if `guest_base` + `mapping.len()` would overflow.
-    pub fn new(mapping: MmapRegion<B>, guest_base: GuestAddress) -> Option<Self> {
-        guest_base
-            .0
-            .checked_add(mapping.size() as u64)
-            .map(|_| Self {
-                mapping,
-                guest_base,
-            })
-    }
-}
-
-#[cfg(not(feature = "xen"))]
-impl<B: NewBitmap> GuestRegionMmap<B> {
-    /// Create a new memory-mapped memory region from guest's physical memory, size and file.
-    pub fn from_range(
-        addr: GuestAddress,
-        size: usize,
-        file: Option<FileOffset>,
-    ) -> result::Result<Self, FromRangesError> {
-        let region = if let Some(ref f_off) = file {
-            MmapRegion::from_file(f_off.clone(), size)?
-        } else {
-            MmapRegion::new(size)?
-        };
-
-        Self::new(region, addr).ok_or(FromRangesError::InvalidGuestRegion)
-    }
-}
-
-#[cfg(feature = "xen")]
-impl<B: NewBitmap> GuestRegionMmap<B> {
-    /// Create a new Unix memory-mapped memory region from guest's physical memory, size and file.
-    /// This must only be used for tests, doctests, benches and is not designed for end consumers.
-    pub fn from_range(
-        addr: GuestAddress,
-        size: usize,
-        file: Option<FileOffset>,
-    ) -> result::Result<Self, FromRangesError> {
-        let range = MmapRange::new_unix(size, file, addr);
-
-        let region = MmapRegion::from_range(range)?;
-        Self::new(region, addr).ok_or(FromRangesError::InvalidGuestRegion)
-    }
-}
-
-impl<B: Bitmap> GuestMemoryRegion for GuestRegionMmap<B> {
-    type B = B;
-
-    fn len(&self) -> GuestUsize {
-        self.mapping.size() as GuestUsize
-    }
-
-    fn start_addr(&self) -> GuestAddress {
-        self.guest_base
-    }
-
-    fn bitmap(&self) -> BS<'_, Self::B> {
-        self.mapping.bitmap().slice_at(0)
-    }
-
-    fn get_host_address(&self, addr: MemoryRegionAddress) -> guest_memory::Result<*mut u8> {
-        // Not sure why wrapping_offset is not unsafe.  Anyway this
-        // is safe because we've just range-checked addr using check_address.
-        self.check_address(addr)
-            .ok_or(guest_memory::Error::InvalidBackendAddress)
-            .map(|addr| {
-                self.mapping
-                    .as_ptr()
-                    .wrapping_offset(addr.raw_value() as isize)
-            })
-    }
-
-    fn file_offset(&self) -> Option<&FileOffset> {
-        self.mapping.file_offset()
-    }
-
-    fn get_slice(
-        &self,
-        offset: MemoryRegionAddress,
-        count: usize,
-    ) -> guest_memory::Result<VolatileSlice<BS<B>>> {
-        let slice = VolatileMemory::get_slice(&self.mapping, offset.raw_value() as usize, count)?;
-        Ok(slice)
-    }
-
-    #[cfg(target_os = "linux")]
-    fn is_hugetlbfs(&self) -> Option<bool> {
-        self.mapping.is_hugetlbfs()
-    }
-}
-
-impl<B: Bitmap> GuestMemoryRegionBytes for GuestRegionMmap<B> {}
-
-/// [`GuestMemory`](trait.GuestMemory.html) implementation that mmaps the guest's memory
-/// in the current process.
-///
-/// Represents the entire physical memory of the guest by tracking all its memory regions.
-/// Each region is an instance of `GuestRegionMmap`, being backed by a mapping in the
-/// virtual address space of the calling process.
-pub type GuestMemoryMmap<B = ()> = GuestRegionCollection<GuestRegionMmap<B>>;
-
-/// Errors that can happen during [`GuestMemoryMmap::from_ranges`] and related functions.
-#[derive(Debug, thiserror::Error)]
-pub enum FromRangesError {
-    /// Error during construction of [`GuestMemoryMmap`]
-    #[error("Error constructing guest region collection: {0}")]
-    Collection(#[from] GuestRegionCollectionError),
-    /// Error while allocating raw mmap region
-    #[error("Error setting up raw memory for guest region: {0}")]
-    MmapRegion(#[from] MmapRegionError),
-    /// A combination of region length and guest address would overflow.
-    #[error("Combination of guest address and region length invalid (would overflow)")]
-    InvalidGuestRegion,
-}
-
-impl<B: NewBitmap> GuestMemoryMmap<B> {
-    /// Creates a container and allocates anonymous memory for guest memory regions.
-    ///
-    /// Valid memory regions are specified as a slice of (Address, Size) tuples sorted by Address.
-    pub fn from_ranges(ranges: &[(GuestAddress, usize)]) -> result::Result<Self, FromRangesError> {
-        Self::from_ranges_with_files(ranges.iter().map(|r| (r.0, r.1, None)))
-    }
-
-    /// Creates a container and allocates anonymous memory for guest memory regions.
-    ///
-    /// Valid memory regions are specified as a sequence of (Address, Size, [`Option<FileOffset>`])
-    /// tuples sorted by Address.
-    pub fn from_ranges_with_files<A, T>(ranges: T) -> result::Result<Self, FromRangesError>
-    where
-        A: Borrow<(GuestAddress, usize, Option<FileOffset>)>,
-        T: IntoIterator<Item = A>,
-    {
-        Self::from_regions(
-            ranges
-                .into_iter()
-                .map(|x| {
-                    GuestRegionMmap::from_range(x.borrow().0, x.borrow().1, x.borrow().2.clone())
-                })
-                .collect::<Result<Vec<_>, _>>()?,
-        )
-        .map_err(Into::into)
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod tests {
     #![allow(clippy::undocumented_unsafe_blocks)]
     extern crate vmm_sys_util;
 
-    use super::*;
-
-    #[cfg(feature = "backend-bitmap")]
-    use crate::bitmap::AtomicBitmap;
-    use crate::{Bytes, GuestMemory, GuestMemoryError};
+    use crate::bitmap::BS;
+    use crate::{
+        guest_memory, Address, Bytes, FileOffset, GuestAddress, GuestMemory, GuestMemoryError,
+        GuestMemoryRegion, GuestMemoryRegionBytes, GuestUsize, MemoryRegionAddress, VolatileMemory,
+        VolatileSlice,
+    };
 
     use std::io::Write;
     use std::sync::Arc;
@@ -259,7 +81,7 @@ pub(crate) mod tests {
                     }
                 }
 
-                fn bitmap(&self) { }
+                fn bitmap(&self) {}
 
                 fn get_host_address(&self, addr: MemoryRegionAddress) -> guest_memory::Result<*mut u8> {
                     match self {
@@ -286,9 +108,9 @@ pub(crate) mod tests {
 
     any_backend! {
         #[cfg(all(windows, feature = "backend-mmap"))]
-        Windows[GuestRegionMmap<()>],
+        Windows[crate::mmap::windows::GuestRegionWindows<()>],
         #[cfg(all(unix, feature = "backend-mmap", not(feature = "xen")))]
-        Mmap[GuestRegionMmap<()>],
+        Mmap[crate::mmap::unix::GuestRegionMmap<()>],
         #[cfg(all(unix, feature = "backend-mmap", feature = "xen"))]
         Xen[crate::mmap::xen::MmapRegion]
     }
@@ -300,13 +122,19 @@ pub(crate) mod tests {
             let mut regions = Vec::new();
             #[cfg(all(windows, feature = "backend-mmap"))]
             regions.push(AnyRegion::Windows(
-                GuestRegionMmap::new(MmapRegion::from_file(f_off.clone(), size).unwrap(), addr)
-                    .unwrap(),
+                crate::mmap::windows::GuestRegionWindows::new(
+                    crate::mmap::windows::MmapRegion::from_file(f_off.clone(), size).unwrap(),
+                    addr,
+                )
+                .unwrap(),
             ));
             #[cfg(all(unix, feature = "backend-mmap", not(feature = "xen")))]
             regions.push(AnyRegion::Mmap(
-                GuestRegionMmap::new(MmapRegion::from_file(f_off.clone(), size).unwrap(), addr)
-                    .unwrap(),
+                crate::mmap::unix::GuestRegionMmap::new(
+                    MmapRegion::from_file(f_off.clone(), size).unwrap(),
+                    addr,
+                )
+                .unwrap(),
             ));
             #[cfg(all(unix, feature = "backend-mmap", feature = "xen"))]
             regions.push(AnyRegion::Xen(
@@ -324,11 +152,16 @@ pub(crate) mod tests {
             };
             #[cfg(all(windows, feature = "backend-mmap"))]
             regions.push(AnyRegion::Windows(
-                GuestRegionMmap::new(MmapRegion::new(size).unwrap(), addr).unwrap(),
+                crate::mmap::windows::GuestRegionWindows::new(
+                    crate::mmap::windows::MmapRegion::new(size).unwrap(),
+                    addr,
+                )
+                .unwrap(),
             ));
             #[cfg(all(unix, feature = "backend-mmap", not(feature = "xen")))]
             regions.push(AnyRegion::Mmap(
-                GuestRegionMmap::new(MmapRegion::new(size).unwrap(), addr).unwrap(),
+                crate::mmap::unix::GuestRegionMmap::new(MmapRegion::new(size).unwrap(), addr)
+                    .unwrap(),
             ));
             #[cfg(all(unix, feature = "backend-mmap", feature = "xen"))]
             regions.push(AnyRegion::Xen(
@@ -778,14 +611,5 @@ pub(crate) mod tests {
                 MemoryRegionAddress(0x1000),
             );
         }
-    }
-
-    #[test]
-    #[cfg(feature = "backend-bitmap")]
-    fn test_dirty_tracking() {
-        crate::bitmap::tests::test_guest_memory_and_region(|| {
-            crate::GuestMemoryMmap::<AtomicBitmap>::from_ranges(&[(GuestAddress(0), 0x1_0000)])
-                .unwrap()
-        });
     }
 }
