+++ src/region.rs
@@ -124,7 +124,7 @@ pub trait GuestMemoryRegion: Bytes<MemoryRegionAddress, E = GuestMemoryError> {
     /// # Examples (uses the `backend-mmap` feature)
     ///
     /// ```
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # {
     /// # use vm_memory::{GuestAddress, MmapRegion, GuestRegionMmap, GuestMemoryRegion};
     /// # use vm_memory::volatile_memory::{VolatileMemory, VolatileSlice, VolatileRef};
@@ -154,7 +154,7 @@ pub trait GuestMemoryRegion: Bytes<MemoryRegionAddress, E = GuestMemoryError> {
     /// # Examples (uses the `backend-mmap` feature)
     ///
     /// ```
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # {
     /// #   use vm_memory::{GuestAddress, GuestMemory, GuestMemoryMmap, GuestRegionMmap};
     /// let addr = GuestAddress(0x1000);
@@ -335,10 +335,10 @@ impl<R: GuestMemoryRegionBytes> Bytes<MemoryRegionAddress> for R {
     /// * Write a slice at guest address 0x1200.
     ///
     /// ```
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # use vm_memory::{Bytes, GuestAddress, GuestMemoryMmap};
     /// #
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # {
     /// # let start_addr = GuestAddress(0x1000);
     /// # let mut gm = GuestMemoryMmap::<()>::from_ranges(&vec![(start_addr, 0x400)])
@@ -361,10 +361,10 @@ impl<R: GuestMemoryRegionBytes> Bytes<MemoryRegionAddress> for R {
     /// * Read a slice of length 16 at guestaddress 0x1200.
     ///
     /// ```
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # use vm_memory::{Bytes, GuestAddress, GuestMemoryMmap};
     /// #
-    /// # #[cfg(feature = "backend-mmap")]
+    /// # #[cfg(all(feature = "backend-mmap", target_family = "unix", not(feature = "xen")))]
     /// # {
     /// # let start_addr = GuestAddress(0x1000);
     /// # let mut gm = GuestMemoryMmap::<()>::from_ranges(&vec![(start_addr, 0x400)])
